refactor: separate globals and input
