+++ include/input.h
@@ -1,0 +1,8 @@
+#ifndef INPUT_H
+#define INPUT_H
+
+void keyboard();
+void clearKeys();
+void handleInput();
+
+#endif
